@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
@@ -15,7 +15,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="XTI_ApiGeneratorApp.Extensions" Version="1.3.34" />
+		<PackageReference Include="XTI_ApiGeneratorApp.Extensions" Version="1.3.35-dev241113134017721" />
 	</ItemGroup>
 
 	<ItemGroup>
