@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="XTI_App.Extensions" Version="1.8.42" />
-		<PackageReference Include="XTI_WebApp.Api" Version="1.8.42" />
+		<PackageReference Include="XTI_App.Extensions" Version="1.8.43-dev241116182821633" />
+		<PackageReference Include="XTI_WebApp.Api" Version="1.8.43-dev241116182821633" />
 	</ItemGroup>
 
 </Project>
