@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Description>App Client for $(XtiAppName) $(XtiVersion)</Description>
@@ -9,10 +9,10 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0" />
-		<PackageReference Include="XTI_App.Api" Version="1.8.42" />
-		<PackageReference Include="XTI_WebAppClient" Version="1.8.42" />
-		<PackageReference Include="XTI_TempLog.Abstractions" Version="1.2.32" />
+		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.0" />
+		<PackageReference Include="XTI_App.Api" Version="1.8.43-dev241116182821633" />
+		<PackageReference Include="XTI_WebAppClient" Version="1.8.43-dev241116182821633" />
+		<PackageReference Include="XTI_TempLog.Abstractions" Version="1.2.33-dev241116170100740" />
 	</ItemGroup>
 
 </Project>
