@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
@@ -18,7 +18,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="XTI_Secrets.Extensions" Version="1.8.10" />
+		<PackageReference Include="XTI_Secrets.Extensions" Version="1.8.11-dev241112213806362" />
 	</ItemGroup>
 
 	<ItemGroup>
