@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<TypeScriptToolsVersion>4.4</TypeScriptToolsVersion>
@@ -11,12 +11,12 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.TypeScript.MSBuild" Version="5.4.4">
+		<PackageReference Include="Microsoft.TypeScript.MSBuild" Version="5.6.2">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="XTI_WebApp.RCL" Version="1.8.42" />
-		<PackageReference Include="XTI_WebApp.Scheduled" Version="1.8.42" />
+		<PackageReference Include="XTI_WebApp.RCL" Version="1.8.43-dev241116182821633" />
+		<PackageReference Include="XTI_WebApp.Scheduled" Version="1.8.43-dev241116182821633" />
 	</ItemGroup>
 
 	<ItemGroup>
