@@ -93,15 +93,15 @@ public sealed class AppRequest
             );
 
     public AppRequestModel ToModel() =>
-        new AppRequestModel
+        new
         (
             ID: ID,
-            SessionID: record.SessionID,
             Path: record.Path,
             ResourceID: record.ResourceID,
             ModifierID: record.ModifierID,
             TimeStarted: record.TimeStarted,
-            TimeEnded: record.TimeEnded
+            TimeEnded: record.TimeEnded,
+            ActualCount: record.ActualCount
         );
 
     public override string ToString() => $"{nameof(AppRequest)} {ID}";
