@@ -1,5 +1,4 @@
-﻿using XTI_HubWebAppApi.Periodic;
-using XTI_HubWebAppApi.PermanentLog;
+﻿using XTI_HubWebAppApi.PermanentLog;
 
 namespace XTI_HubWebAppApi;
 
@@ -8,5 +7,6 @@ internal static class PermanentLogExtensions
     public static void AddPermanentLogGroupServices(this IServiceCollection services)
     {
         services.AddScoped<LogBatchAction>();
+        services.AddScoped<LogSessionDetailsAction>();
     }
 }
