@@ -3,16 +3,18 @@
 public sealed record AppRequestModel
 (
     int ID,
-    int SessionID,
     string Path,
     int ResourceID,
     int ModifierID,
     DateTimeOffset TimeStarted,
-    DateTimeOffset TimeEnded
+    DateTimeOffset TimeEnded,
+    int ActualCount
 )
 {
     public AppRequestModel()
-        : this(0, 0, "", 0, 0, DateTimeOffset.MaxValue, DateTimeOffset.MaxValue)
+        : this(0, "", 0, 0, DateTimeOffset.MaxValue, DateTimeOffset.MaxValue, 0)
     {
     }
+
+    public bool IsFound() => ID > 0;
 }
