@@ -5,12 +5,14 @@ public sealed record AppSessionModel
     int ID,
     DateTimeOffset TimeStarted,
     DateTimeOffset TimeEnded,
+    string SessionKey,
+    string RequesterKey,
     string RemoteAddress,
     string UserAgent
 )
 {
     public AppSessionModel()
-        :this(0,DateTimeOffset.MaxValue, DateTimeOffset.MaxValue, "", "")
+        :this(0,DateTimeOffset.MaxValue, DateTimeOffset.MaxValue, "", "", "", "")
     {
     }
 }
