@@ -1,15 +1,10 @@
-﻿using DocumentFormat.OpenXml.Office2016.Drawing.ChartDrawing;
-using HubWebApp.Extensions;
-using Microsoft.EntityFrameworkCore;
+﻿using HubWebApp.Extensions;
 using Microsoft.Extensions.Hosting;
 using XTI_App.Extensions;
 using XTI_Core;
 using XTI_Core.Extensions;
 using XTI_Core.Fakes;
-using XTI_Hub.Abstractions;
-using XTI_HubDB.EF;
 using XTI_HubDB.Extensions;
-using XTI_HubWebAppApi.PermanentLog;
 using XTI_WebApp.Abstractions;
 using XTI_WebApp.Api;
 
@@ -45,7 +40,7 @@ internal sealed class HubTestHost
         builder.Services.AddSingleton(sp => sp.GetRequiredService<DefaultWebAppOptions>().DB);
         builder.Services.AddHubDbContextForSqlServer();
         builder.Services.AddScoped<HubFactory>();
-        builder.Services.AddScoped<PermanentLog>();
+        builder.Services.AddScoped<EfPermanentLog>();
         builder.Services.AddScoped<AppFromPath>();
         builder.Services.AddScoped<IHashedPasswordFactory, Md5HashedPasswordFactory>();
         builder.Services.AddScoped<AccessForAuthenticate, JwtAccess>();
