@@ -4,14 +4,15 @@ using Microsoft.Extensions.Hosting;
 using XTI_App.Abstractions;
 using XTI_App.Api;
 using XTI_Core;
-using XTI_HubAppClient;
+using XTI_Hub;
 using XTI_HubAppClient.ServiceApp.Extensions;
 using XTI_PermanentLog;
+using XTI_PermanentLog.Implementations;
 using XTI_Schedule;
 using XTI_SupportServiceAppApi;
 using XTI_TempLog;
-using XTI_TempLog.Abstractions;
 using XTI_TempLog.Extensions;
+using XTI_HubDB.Extensions;
 
 var hostBuilder = XtiServiceAppHost.CreateDefault(SupportInfo.AppKey, args)
     .ConfigureServices((hostContext, services) =>
@@ -19,15 +20,36 @@ var hostBuilder = XtiServiceAppHost.CreateDefault(SupportInfo.AppKey, args)
         services.AddSupportAppApiServices();
         services.AddScoped<AppApiFactory, SupportAppApiFactory>();
         services.AddScoped(sp => (SupportAppApi)sp.GetRequiredService<IAppApi>());
-        services.AddScoped<ITempLogs>(sp =>
+        services.AddScoped<ITempLogsV1>(sp =>
         {
             var dataProtector = sp.GetDataProtector("XTI_TempLog");
             var appKey = sp.GetRequiredService<AppKey>();
             var appDataFolder = sp.GetRequiredService<XtiFolder>().AppDataFolder();
-            return new DiskTempLogs(dataProtector, appDataFolder.Path(), "TempLogs");
+            return new DiskTempLogsV1(dataProtector, appDataFolder.Path(), "TempLogs");
         });
-        services.AddScoped<IPermanentLogClient, PermanentLogClient>();
+        services.AddHubDbContextForSqlServer();
+        services.AddScoped<HubFactory>();
+        services.AddScoped<EfPermanentLog>();
+        services.AddScoped<HcPermanentLog>();
+        services.AddScoped
+        (
+            sp =>
+            {
+                IPermanentLog permanentLog;
+                var options = sp.GetRequiredService<SupportServiceAppOptions>();
+                if (options.PermanentLogType == "DB")
+                {
+                    permanentLog = sp.GetRequiredService<EfPermanentLog>();
+                }
+                else
+                {
+                    permanentLog = sp.GetRequiredService<HcPermanentLog>();
+                }
+                return permanentLog;
+            }
+        );
         services.AddScoped<TempToPermanentLog>();
+        services.AddScoped<TempToPermanentLogV1>();
         services.AddAppAgenda
         (
             (sp, agenda) =>
@@ -37,6 +59,18 @@ var hostBuilder = XtiServiceAppHost.CreateDefault(SupportInfo.AppKey, args)
                     (api, agendaItem) =>
                     {
                         agendaItem.Action(api.PermanentLog.MoveToPermanent)
+                            .Interval(TimeSpan.FromMinutes(2))
+                            .AddSchedule
+                            (
+                                Schedule.EveryDay().At(TimeRange.AllDay())
+                            );
+                    }
+                );
+                agenda.AddScheduled<SupportAppApi>
+                (
+                    (api, agendaItem) =>
+                    {
+                        agendaItem.Action(api.PermanentLog.MoveToPermanentV1)
                             .Interval(TimeSpan.FromMinutes(5))
                             .AddSchedule
                             (
@@ -49,6 +83,19 @@ var hostBuilder = XtiServiceAppHost.CreateDefault(SupportInfo.AppKey, args)
                     (api, agendaItem) =>
                     {
                         agendaItem.Action(api.PermanentLog.Retry)
+                            .DelayAfterStart(TimeSpan.FromMinutes(1))
+                            .Interval(TimeSpan.FromHours(1))
+                            .AddSchedule
+                            (
+                                Schedule.EveryDay().At(TimeRange.AllDay())
+                            );
+                    }
+                );
+                agenda.AddScheduled<SupportAppApi>
+                (
+                    (api, agendaItem) =>
+                    {
+                        agendaItem.Action(api.PermanentLog.RetryV1)
                             .DelayAfterStart(TimeSpan.FromMinutes(1))
                             .Interval(TimeSpan.FromHours(1))
                             .AddSchedule
@@ -76,6 +123,9 @@ var hostBuilder = XtiServiceAppHost.CreateDefault(SupportInfo.AppKey, args)
             (api, throttledLogs) =>
             {
                 throttledLogs.Throttle(api.PermanentLog.MoveToPermanent)
+                    .Requests().ForOneHour()
+                    .Exceptions().For(5).Minutes();
+                throttledLogs.Throttle(api.PermanentLog.MoveToPermanentV1)
                     .Requests().ForOneHour()
                     .Exceptions().For(5).Minutes();
             }
